backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
click 
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
fastapi==0.115.11
h11 
idna==3.10
kombu==5.4.2
mongoengine==0.29.1
prompt_toolkit==3.0.50
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
watchdog
openai
pydantic
reportlab
google-search-results
python-multipart
websockets